Beta Version
For this push, I worked mostly on the first two screens. The intro sequence now has a border of images that are pulled from a JSON. The radio page now has a part that reacts to the amplitude of the audio. Also changed the font and added a "widget" of sorts that displays different names of internees and where they were born.